DAT-901/T-DAT-901-PAR_10/CryptoViz_Back/requirements.txt
absl-py
annotated-types
anyio
asgiref
astunparse
bcrypt
black
cachetools
certifi
cfgv
charset-normalizer
click
colorama
contourpy
cycler
dash
dash-core-components
dash-html-components
dash-table
distlib
Django
django-base64field
django-cors-headers
djangorestframework
dnspython
duckdb
ecdsa
email-validator
fastapi
filelock
Flask
flatbuffers
fonttools
gast
google-auth
google-auth-oauthlib
google-pasta
greenlet
grpcio
h11
h5py
httpcore
httpx
identify
idna
iniconfig
itsdangerous
jax
Jinja2
keras
kiwisolver
libclang
Markdown
MarkupSafe
matplotlib
ml-dtypes
mnist-dnn-api
mpmath
mypy-extensions
networkx
nodeenv
numpy
oauthlib
opencv-python
opt-einsum
packaging
pandas
passlib
pathspec
Pillow
platformdirs
plotly
pluggy
polars
pre-commit
protobuf
psycopg2-binary
pyarrow
pyasn1
pyasn1-modules
pydantic
pydantic_core
PyJWT
pyparsing
pytest
pytest-asyncio
python-dateutil
python-jose
python-mnist
python-multipart
pytz
PyYAML
requests
requests-oauthlib
rsa
scipy
six
sniffio
SQLAlchemy
sqlparse
starlette
sympy
tenacity
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
torch
torchaudio
torchvision
typing_extensions
tzdata
urllib3
uvicorn
virtualenv
Werkzeug
wrapt
